python/paddlecloud/requirements.txt
django-user-accounts==2.0.1
pinax-webanalytics==3.0.0
pinax-eventlog==1.1.2

Django==1.11
pinax-theme-bootstrap==8.0.1

kubernetes==2.0
MySQL-python
djangorestframework
django_sendmail_backend
nodejs
